requirements.txt
ultralytics
opencv-python
requests
PyYAML
numpy
torch
onnx
streamlit